+++ layers/GatoGraphQLForWP/plugins/gatographql/blocks/endpoint-voyager/webpack.config.js
@@ -73,7 +73,8 @@ config.module.rules.push(
  * @see https://webpack.js.org/configuration/output/#outputjsonpfunction
  */
 // ------------------------------------------------------
-config.output.jsonpFunction = 'webpackJsonpGatoGraphQLEndpointVoyager';
+// Needed for webpack v4, so commented out for webpack v5
+// config.output.jsonpFunction = 'webpackJsonpGatoGraphQLEndpointVoyager';
 // ------------------------------------------------------
 
 module.exports = config;