+++ layers/GatoGraphQLForWP/plugins/gatographql/blocks/persisted-query-endpoint-options/webpack.config.js
@@ -94,7 +94,8 @@ config.module.rules.push(
  * @see https://webpack.js.org/configuration/output/#outputjsonpfunction
  */
 // ------------------------------------------------------
-config.output.jsonpFunction = 'webpackJsonpGatoGraphQLPersistedQueryEndpointOptions';
+// Needed for webpack v4, so commented out for webpack v5
+// config.output.jsonpFunction = 'webpackJsonpGatoGraphQLPersistedQueryEndpointOptions';
 // ------------------------------------------------------
 
 module.exports = config;