+++ layers/GatoGraphQLForWP/plugins/gatographql/blocks/schema-config-global-fields/webpack.config.js
@@ -66,7 +66,8 @@ config.module.rules.push(
  * @see https://webpack.js.org/configuration/output/#outputjsonpfunction
  */
 // ------------------------------------------------------
-config.output.jsonpFunction = 'webpackJsonpGatoGraphQLSchemaConfigGlobalFields';
+// Needed for webpack v4, so commented out for webpack v5
+// config.output.jsonpFunction = 'webpackJsonpGatoGraphQLSchemaConfigGlobalFields';
 // ------------------------------------------------------
 
 module.exports = config;