+++ layers/GatoGraphQLForWP/plugins/gatographql/blocks/schema-config-schema-settings/webpack.config.js
@@ -67,7 +67,8 @@ config.module.rules.push(
  * @see https://webpack.js.org/configuration/output/#outputjsonpfunction
  */
 // ------------------------------------------------------
-config.output.jsonpFunction = 'webpackJsonpGatographqlSchemaConfigSchemaSettings';
+// Needed for webpack v4, so commented out for webpack v5
+// config.output.jsonpFunction = 'webpackJsonpGatographqlSchemaConfigSchemaSettings';
 // ------------------------------------------------------
 
 module.exports = config;