+++ layers/GatoGraphQLForWP/plugins/gatographql/editor-scripts/persisted-query-endpoint-overview/webpack.config.js
@@ -15,7 +15,8 @@ config.module.rules.push(
  * @see https://webpack.js.org/configuration/output/#outputjsonpfunction
  */
 // ------------------------------------------------------
-config.output.jsonpFunction = 'webpackJsonpGatoGraphQLPersistedQueryEndpointOverview';
+// Needed for webpack v4, so commented out for webpack v5
+// config.output.jsonpFunction = 'webpackJsonpGatoGraphQLPersistedQueryEndpointOverview';
 // ------------------------------------------------------
 
 module.exports = config;