+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -4,7 +4,18 @@ All notable changes to `gatographql` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## 2.6.0 - DATE
+## 2.7.0 - DATE
+
+### Improvements
+
+- Added compatibility with WordPress 6.6 (#2717)
+
+### Fixed
+
+- Catch exception if dependency version is not semver (#2712)
+- Convert entries in JSON dictionary of variables in persisted query from array to object (#2715)
+
+## 2.6.0 - 01/07/2024
 
 ### Improvements